requirements.txt
#  The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

apscheduler>=3.5.1 # MIT License
keystoneauth1>=3.4.0 # Apache-2.0
keystonemiddleware>=4.21.0 # Apache-2.0
os-resource-classes>=0.4.0
oslo.concurrency>=3.26.0 # Apache-2.0
oslo.cache>=1.29.0 # Apache-2.0
oslo.config>=6.8.0 # Apache-2.0
oslo.context>=2.21.0 # Apache-2.0
oslo.db>=4.44.0 # Apache-2.0
oslo.i18n>=3.20.0 # Apache-2.0
oslo.log>=3.37.0 # Apache-2.0
oslo.messaging>=8.1.2 # Apache-2.0
oslo.policy>=3.6.0 # Apache-2.0
oslo.reports>=1.27.0 # Apache-2.0
oslo.serialization>=2.25.0 # Apache-2.0
oslo.service>=1.30.0 # Apache-2.0
oslo.utils>=3.36.0 # Apache-2.0
oslo.versionedobjects>=1.32.0 # Apache-2.0
pbr>=3.1.1 # Apache-2.0
pecan>=1.3.2 # BSD
python-glanceclient>=2.9.1 # Apache-2.0
python-keystoneclient>=3.15.0 # Apache-2.0
python-openstackclient>=3.14.0 # Apache-2.0
SQLAlchemy>=1.2.5 # MIT
microversion_parse>=0.2.1 # Apache-2.0
futurist>=1.8.0 # Apache-2.0